+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- 🐛 fix: sync.Mutex added to prevent unwanted behaviors
 
 ## [v0.0.1] - 2022-12-08